support PCIeX8 for vc709
